Rename automation to integration
